Valider innsendt skjema på frontend server.
/api/pdf handler vil no validere innsendte data på samme måte som klient gjere før forespørsel går til server. For å sikre at validering faktisk er korrekt utført på innsendte data.
