+++ src/app/(withApis)/SelfApiContext.ts
@@ -1,9 +1,9 @@
 import { createContext } from "react";
 import { SelfApi } from "@/integrations/self/SelfApi";
-import { PSBLegeerklæringInnsending } from "@/integrations/helseopplysningerserver/types/HelseopplysningerTypes";
+import LegeerklaeringDokument from "@/models/LegeerklaeringDokument";
 
 const uninitialized: SelfApi = {
-    generatePdf(innsending: PSBLegeerklæringInnsending): Promise<Blob> {
+    generatePdf(innsending: LegeerklaeringDokument): Promise<Blob> {
         throw new Error("SelfApiContext not set")
     }
 }
