+++ src/app/api/pdf/route.ts
@@ -6,6 +6,11 @@ import HelseopplysningerClient from "@/integrations/helseopplysningerserver/Hels
 import { PSBLegeerklæringInnsending } from "@/integrations/helseopplysningerserver/types/HelseopplysningerTypes";
 import { FhirSession } from "@/auth/fhir/FhirSession";
 import { JwtVerificationInput } from "@/auth/fhir/JwtVerificationInput";
+import { legeerklaeringDokumentSchema } from "@/app/components/legeerklaering/legeerklaeringDokumentSchema";
+import {
+    mapTilPSBLegeerklæringInnsending
+} from "@/integrations/helseopplysningerserver/types/mapTilPSBLegeerklæringInnsending";
+import { ValidationError } from "yup";
 
 const initHelseopplysningerApi = () => process.env.FAKE_HELSEOPPLYSNINGER === "fake1" ?
     new FakeHelseopplysningerApi1() :
@@ -17,16 +22,27 @@ export const POST = async (request: NextRequest): Promise<Response> => {
     const authHeader = request.headers.get("Authorization") || ""
     const fhirSession = await FhirSession.fromVerifiedJWT(JwtVerificationInput.fromAuthorizationHeader(authHeader))
     logger.info("genererer legeerklæring pdf...");
-    const innsending = await request.json() as PSBLegeerklæringInnsending;
-    // Check that hpr-number of session equals posted document
-    if(innsending.lege.hpr === fhirSession.hprNumber) {
-        // XXX Her kunne vi evt ha validering av nokre innsendte skjemaopplysninger opp mot fhir api, før registrering av
-        // data i NAV sine system. Feks at lege sitt namn er korrekt, spesialiststatus, pasient-namn, etc.
-        const helseopplysningerApi = initHelseopplysningerApi()
-        const pdfBlob = await helseopplysningerApi.generatePdf(innsending)
-        return new Response(pdfBlob)
+    // Valider innsendte skjemadata
+    const requestJson = await request.json()
+    try {
+        const legeerklæringDokument = await legeerklaeringDokumentSchema.validate(requestJson)
+        const innsending = mapTilPSBLegeerklæringInnsending(legeerklæringDokument)
+        // Check that hpr-number of session equals posted document
+        if(innsending.lege.hpr === fhirSession.hprNumber) {
+            // XXX Her kunne vi evt ha validering av nokre innsendte skjemaopplysninger opp mot fhir api, før registrering av
+            // data i NAV sine system. Feks at lege sitt namn er korrekt, spesialiststatus, pasient-namn, etc.
+            const helseopplysningerApi = initHelseopplysningerApi()
+            const pdfBlob = await helseopplysningerApi.generatePdf(innsending)
+            return new Response(pdfBlob)
 
-    } else {
-        return new NextResponse(`fhir session hpr-nummer ulikt legeerklæring (${fhirSession.hprNumber} != ${innsending.lege.hpr})`, {status: 400})
+        } else {
+            return new NextResponse(`fhir session hpr-nummer ulikt legeerklæring (${fhirSession.hprNumber} != ${innsending.lege.hpr})`, {status: 400})
+        }
+    } catch(e) {
+        if(e instanceof ValidationError) {
+            return new NextResponse(`${e.message}`, {status: 400})
+        } else {
+            throw e
+        }
     }
 }
